requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
AnyQt==0.1.0
asttokens==2.0.5
asyncpg==0.25.0
autopep8==1.6.0
backcall==0.2.0
baycomp==1.0.2
Bottleneck==1.3.4
CacheControl==0.12.11
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
databases==0.5.5
debugpy==1.6.0
decorator==5.1.1
dictdiffer==0.9.0
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastapi==0.74.1
fonttools==4.33.3
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.7
httpx==0.22.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
iso8601==0.1.16
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.0
jupyter-core==4.10.0
keyring==23.5.0
keyrings.alt==4.1.0
kiwisolver==1.4.2
lockfile==0.12.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
msgpack==1.0.3
nest-asyncio==1.5.5
networkx==2.8
numpy==1.22.2
openpyxl==3.0.9
openTSNE==0.6.2
orange-canvas-core==0.1.26
orange-widget-base==4.17.0
Orange3==3.32.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
priority==2.0.0
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pycodestyle==2.8.0
pydantic==1.9.0
Pygments==2.12.0
pyparsing==3.0.8
pypika-tortoise==0.1.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqtgraph==0.12.3
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-louvain==0.16
pytz==2021.3
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==22.3.0
qasync==0.23.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.8.0
serverfiles==0.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
stack-data==0.2.0
starlette==0.17.1
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tortoise-orm==0.18.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.18.1
wcwidth==0.2.5
wsproto==1.0.0
xlrd==2.0.1
XlsxWriter==3.0.3
zipp==3.8.0
